third_party/blink/web_tests/virtual/third-party-storage-partitioning/external/wpt/service-workers/service-worker/partitioned-claim.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] ServiceWorker's clients.claim() is partitioned
  assert_equals: 3p iframe was successfully claimed expected "success" but got "failure"
Harness: the test ran to completion.
